lingvo/requirements.txt
numpy
scipy
lingvo
tensorflow-text
librosa
colortimelog
